refactor[CountryInfoManager]: renamed component and fix item mapping
