+++ src/app/components/CountryInfoCard/CountryInfoManager.module.scss
@@ -10,6 +10,10 @@
   border-radius: 1rem;
   padding: 1rem;
   overflow: auto;
+  
+}
+
+.imageWrapper {
   cursor: pointer;
 }
